Refactor SA code, implement read-only gRPC interface
(cherry picked from commit c5a6328682819f1dadd63a8039d65d18b4da6a84)
