+++ sa/sa.go
@@ -2,24 +2,17 @@ package sa
 
 import (
 	"context"
-	"crypto/sha256"
 	"crypto/x509"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"math/big"
-	"net"
-	"regexp"
 	"strings"
-	"sync"
 	"time"
 
 	"github.com/jmhodges/clock"
 	"github.com/prometheus/client_golang/prometheus"
 	"golang.org/x/crypto/ocsp"
 	"google.golang.org/protobuf/types/known/emptypb"
-	"google.golang.org/protobuf/types/known/timestamppb"
-	jose "gopkg.in/square/go-jose.v2"
 
 	"github.com/letsencrypt/boulder/core"
 	corepb "github.com/letsencrypt/boulder/core/proto"
@@ -27,41 +20,21 @@ import (
 	berrors "github.com/letsencrypt/boulder/errors"
 	"github.com/letsencrypt/boulder/features"
 	bgrpc "github.com/letsencrypt/boulder/grpc"
-	"github.com/letsencrypt/boulder/identifier"
 	blog "github.com/letsencrypt/boulder/log"
 	"github.com/letsencrypt/boulder/revocation"
-	rocsp_config "github.com/letsencrypt/boulder/rocsp/config"
 	sapb "github.com/letsencrypt/boulder/sa/proto"
 )
 
 var (
-	errIncompleteRequest     = errors.New("incomplete gRPC request message")
-	validIncidentTableRegexp = regexp.MustCompile(`^incident_[0-9a-zA-Z_]{1,100}$`)
+	errIncompleteRequest = errors.New("incomplete gRPC request message")
 )
-
-type certCountFunc func(db db.Selector, domain string, timeRange *sapb.Range) (int64, time.Time, error)
 
 // SQLStorageAuthority defines a Storage Authority
 type SQLStorageAuthority struct {
 	sapb.UnimplementedStorageAuthorityServer
-	dbMap          *db.WrappedMap
-	dbReadOnlyMap  *db.WrappedMap
-	dbIncidentsMap *db.WrappedMap
+	*SQLStorageAuthorityRO
 
-	// Short issuer map used by rocsp.
-	shortIssuers []rocsp_config.ShortIDIssuer
-
-	clk clock.Clock
-	log blog.Logger
-
-	// For RPCs that generate multiple, parallelizable SQL queries, this is the
-	// max parallelism they will use (to avoid consuming too many MariaDB
-	// threads).
-	parallelismPerRPC int
-
-	// We use function types here so we can mock out this internal function in
-	// unittests.
-	countCertificatesByName certCountFunc
+	dbMap *db.WrappedMap
 
 	// rateLimitWriteErrors is a Counter for the number of times
 	// a ratelimit update transaction failed during AddCertificate request
@@ -69,37 +42,16 @@ type SQLStorageAuthority struct {
 	// transactions fail and so use this stat to maintain visibility into the rate
 	// this occurs.
 	rateLimitWriteErrors prometheus.Counter
-
-	// redisStoreResponse is a counter of OCSP responses written to redis by
-	// result.
-	redisStoreResponse *prometheus.CounterVec
 }
 
-// orderFQDNSet contains the SHA256 hash of the lowercased, comma joined names
-// from a new-order request, along with the corresponding orderID, the
-// registration ID, and the order expiry. This is used to find
-// existing orders for reuse.
-type orderFQDNSet struct {
-	ID             int64
-	SetHash        []byte
-	OrderID        int64
-	RegistrationID int64
-	Expires        time.Time
-}
-
-// NewSQLStorageAuthority provides persistence using a SQL backend for
-// Boulder. It will modify the given gorp.DbMap by adding relevant tables.
-func NewSQLStorageAuthority(
+// NewSQLStorageAuthorityWrapping provides persistence using a SQL backend for
+// Boulder. It takes a read-only storage authority to wrap.
+func NewSQLStorageAuthorityWrapping(
+	ssaro *SQLStorageAuthorityRO,
 	dbMap *db.WrappedMap,
-	dbReadOnlyMap *db.WrappedMap,
-	dbIncidentsMap *db.WrappedMap,
-	shortIssuers []rocsp_config.ShortIDIssuer,
-	clk clock.Clock,
-	logger blog.Logger,
 	stats prometheus.Registerer,
-	parallelismPerRPC int,
 ) (*SQLStorageAuthority, error) {
-	SetSQLDebug(dbMap, logger)
+	SetSQLDebug(dbMap, ssaro.log)
 
 	rateLimitWriteErrors := prometheus.NewCounter(prometheus.CounterOpts{
 		Name: "rate_limit_write_errors",
@@ -107,330 +59,32 @@ func NewSQLStorageAuthority(
 	})
 	stats.MustRegister(rateLimitWriteErrors)
 
-	redisStoreResponse := prometheus.NewCounterVec(prometheus.CounterOpts{
-		Name: "redis_store_response",
-		Help: "Count of OCSP Response writes to redis",
-	}, []string{"result"})
-	stats.MustRegister(redisStoreResponse)
-
 	ssa := &SQLStorageAuthority{
-		dbMap:                dbMap,
-		dbReadOnlyMap:        dbReadOnlyMap,
-		dbIncidentsMap:       dbIncidentsMap,
-		shortIssuers:         shortIssuers,
-		clk:                  clk,
-		log:                  logger,
-		parallelismPerRPC:    parallelismPerRPC,
-		rateLimitWriteErrors: rateLimitWriteErrors,
-		redisStoreResponse:   redisStoreResponse,
+		SQLStorageAuthorityRO: ssaro,
+		dbMap:                 dbMap,
+		rateLimitWriteErrors:  rateLimitWriteErrors,
 	}
-
-	ssa.countCertificatesByName = ssa.countCertificates
 
 	return ssa, nil
 }
 
-// GetRegistration obtains a Registration by ID
-func (ssa *SQLStorageAuthority) GetRegistration(ctx context.Context, req *sapb.RegistrationID) (*corepb.Registration, error) {
-	if req == nil || req.Id == 0 {
-		return nil, errIncompleteRequest
-	}
-
-	const query = "WHERE id = ?"
-	model, err := selectRegistration(ssa.dbMap.WithContext(ctx), query, req.Id)
-	if err != nil {
-		if db.IsNoRows(err) {
-			return nil, berrors.NotFoundError("registration with ID '%d' not found", req.Id)
-		}
-		return nil, err
-	}
-
-	return registrationModelToPb(model)
-}
-
-// GetRegistrationByKey obtains a Registration by JWK
-func (ssa *SQLStorageAuthority) GetRegistrationByKey(ctx context.Context, req *sapb.JSONWebKey) (*corepb.Registration, error) {
-	if req == nil || len(req.Jwk) == 0 {
-		return nil, errIncompleteRequest
-	}
-
-	var jwk jose.JSONWebKey
-	err := jwk.UnmarshalJSON(req.Jwk)
+// NewSQLStorageAuthority provides persistence using a SQL backend for
+// Boulder. It constructs its own read-only storage authority to wrap.
+func NewSQLStorageAuthority(
+	dbMap *db.WrappedMap,
+	dbReadOnlyMap *db.WrappedMap,
+	dbIncidentsMap *db.WrappedMap,
+	parallelismPerRPC int,
+	clk clock.Clock,
+	logger blog.Logger,
+	stats prometheus.Registerer,
+) (*SQLStorageAuthority, error) {
+	ssaro, err := NewSQLStorageAuthorityRO(dbReadOnlyMap, dbIncidentsMap, parallelismPerRPC, clk, logger)
 	if err != nil {
 		return nil, err
 	}
 
-	const query = "WHERE jwk_sha256 = ?"
-	sha, err := core.KeyDigestB64(jwk.Key)
-	if err != nil {
-		return nil, err
-	}
-	model, err := selectRegistration(ssa.dbMap.WithContext(ctx), query, sha)
-	if err != nil {
-		if db.IsNoRows(err) {
-			return nil, berrors.NotFoundError("no registrations with public key sha256 %q", sha)
-		}
-		return nil, err
-	}
-
-	return registrationModelToPb(model)
-}
-
-// incrementIP returns a copy of `ip` incremented at a bit index `index`,
-// or in other words the first IP of the next highest subnet given a mask of
-// length `index`.
-// In order to easily account for overflow, we treat ip as a big.Int and add to
-// it. If the increment overflows the max size of a net.IP, return the highest
-// possible net.IP.
-func incrementIP(ip net.IP, index int) net.IP {
-	bigInt := new(big.Int)
-	bigInt.SetBytes([]byte(ip))
-	incr := new(big.Int).Lsh(big.NewInt(1), 128-uint(index))
-	bigInt.Add(bigInt, incr)
-	// bigInt.Bytes can be shorter than 16 bytes, so stick it into a
-	// full-sized net.IP.
-	resultBytes := bigInt.Bytes()
-	if len(resultBytes) > 16 {
-		return net.ParseIP("ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff")
-	}
-	result := make(net.IP, 16)
-	copy(result[16-len(resultBytes):], resultBytes)
-	return result
-}
-
-// ipRange returns a range of IP addresses suitable for querying MySQL for the
-// purpose of rate limiting using a range that is inclusive on the lower end and
-// exclusive at the higher end. If ip is an IPv4 address, it returns that address,
-// plus the one immediately higher than it. If ip is an IPv6 address, it applies
-// a /48 mask to it and returns the lowest IP in the resulting network, and the
-// first IP outside of the resulting network.
-func ipRange(ip net.IP) (net.IP, net.IP) {
-	ip = ip.To16()
-	// For IPv6, match on a certain subnet range, since one person can commonly
-	// have an entire /48 to themselves.
-	maskLength := 48
-	// For IPv4 addresses, do a match on exact address, so begin = ip and end =
-	// next higher IP.
-	if ip.To4() != nil {
-		maskLength = 128
-	}
-
-	mask := net.CIDRMask(maskLength, 128)
-	begin := ip.Mask(mask)
-	end := incrementIP(begin, maskLength)
-
-	return begin, end
-}
-
-// CountRegistrationsByIP returns the number of registrations created in the
-// time range for a single IP address.
-func (ssa *SQLStorageAuthority) CountRegistrationsByIP(ctx context.Context, req *sapb.CountRegistrationsByIPRequest) (*sapb.Count, error) {
-	if len(req.Ip) == 0 || req.Range.Earliest == 0 || req.Range.Latest == 0 {
-		return nil, errIncompleteRequest
-	}
-
-	var count int64
-	err := ssa.dbReadOnlyMap.WithContext(ctx).SelectOne(
-		&count,
-		`SELECT COUNT(1) FROM registrations
-		 WHERE
-		 initialIP = :ip AND
-		 :earliest < createdAt AND
-		 createdAt <= :latest`,
-		map[string]interface{}{
-			"ip":       req.Ip,
-			"earliest": time.Unix(0, req.Range.Earliest),
-			"latest":   time.Unix(0, req.Range.Latest),
-		})
-	if err != nil {
-		return &sapb.Count{Count: -1}, err
-	}
-	return &sapb.Count{Count: count}, nil
-}
-
-// CountRegistrationsByIPRange returns the number of registrations created in
-// the time range in an IP range. For IPv4 addresses, that range is limited to
-// the single IP. For IPv6 addresses, that range is a /48, since it's not
-// uncommon for one person to have a /48 to themselves.
-func (ssa *SQLStorageAuthority) CountRegistrationsByIPRange(ctx context.Context, req *sapb.CountRegistrationsByIPRequest) (*sapb.Count, error) {
-	if len(req.Ip) == 0 || req.Range.Earliest == 0 || req.Range.Latest == 0 {
-		return nil, errIncompleteRequest
-	}
-
-	var count int64
-	beginIP, endIP := ipRange(req.Ip)
-	err := ssa.dbReadOnlyMap.WithContext(ctx).SelectOne(
-		&count,
-		`SELECT COUNT(1) FROM registrations
-		 WHERE
-		 :beginIP <= initialIP AND
-		 initialIP < :endIP AND
-		 :earliest < createdAt AND
-		 createdAt <= :latest`,
-		map[string]interface{}{
-			"earliest": time.Unix(0, req.Range.Earliest),
-			"latest":   time.Unix(0, req.Range.Latest),
-			"beginIP":  beginIP,
-			"endIP":    endIP,
-		})
-	if err != nil {
-		return &sapb.Count{Count: -1}, err
-	}
-	return &sapb.Count{Count: count}, nil
-}
-
-// CountCertificatesByNames counts, for each input domain, the number of
-// certificates issued in the given time range for that domain and its
-// subdomains. It returns a map from domains to counts and a timestamp. The map
-// of domains to counts is guaranteed to contain an entry for each input domain,
-// so long as err is nil. The timestamp is the earliest time a certificate was
-// issued for any of the domains during the provided range of time. Queries will
-// be run in parallel. If any of them error, only one error will be returned.
-func (ssa *SQLStorageAuthority) CountCertificatesByNames(ctx context.Context, req *sapb.CountCertificatesByNamesRequest) (*sapb.CountByNames, error) {
-	if len(req.Names) == 0 || req.Range.Earliest == 0 || req.Range.Latest == 0 {
-		return nil, errIncompleteRequest
-	}
-
-	work := make(chan string, len(req.Names))
-	type result struct {
-		err      error
-		count    int64
-		earliest time.Time
-		domain   string
-	}
-	results := make(chan result, len(req.Names))
-	for _, domain := range req.Names {
-		work <- domain
-	}
-	close(work)
-	var wg sync.WaitGroup
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-	// We may perform up to 100 queries, depending on what's in the certificate
-	// request. Parallelize them so we don't hit our timeout, but limit the
-	// parallelism so we don't consume too many threads on the database.
-	for i := 0; i < ssa.parallelismPerRPC; i++ {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			for domain := range work {
-				select {
-				case <-ctx.Done():
-					results <- result{err: ctx.Err()}
-					return
-				default:
-				}
-				count, earliest, err := ssa.countCertificatesByName(ssa.dbReadOnlyMap.WithContext(ctx), domain, req.Range)
-				if err != nil {
-					results <- result{err: err}
-					// Skip any further work
-					cancel()
-					return
-				}
-				results <- result{
-					count:    count,
-					earliest: earliest,
-					domain:   domain,
-				}
-			}
-		}()
-	}
-	wg.Wait()
-	close(results)
-
-	// Set earliest to the latest possible time, so that we can find the
-	// earliest certificate in the results.
-	earliest := timestamppb.New(time.Unix(0, req.Range.Latest))
-	counts := make(map[string]int64)
-	for r := range results {
-		if r.err != nil {
-			return nil, r.err
-		}
-		counts[r.domain] = r.count
-		if !r.earliest.IsZero() && r.earliest.Before(earliest.AsTime()) {
-			earliest = timestamppb.New(r.earliest)
-		}
-	}
-
-	// If we didn't find any certificates in the range, earliest should be set
-	// to a zero value.
-	if len(counts) == 0 {
-		earliest = &timestamppb.Timestamp{}
-	}
-	return &sapb.CountByNames{Counts: counts, Earliest: earliest}, nil
-}
-
-func ReverseName(domain string) string {
-	labels := strings.Split(domain, ".")
-	for i, j := 0, len(labels)-1; i < j; i, j = i+1, j-1 {
-		labels[i], labels[j] = labels[j], labels[i]
-	}
-	return strings.Join(labels, ".")
-}
-
-// GetCertificate takes a serial number and returns the corresponding
-// certificate, or error if it does not exist.
-func (ssa *SQLStorageAuthority) GetCertificate(ctx context.Context, req *sapb.Serial) (*corepb.Certificate, error) {
-	if req == nil || req.Serial == "" {
-		return nil, errIncompleteRequest
-	}
-	if !core.ValidSerial(req.Serial) {
-		return nil, fmt.Errorf("Invalid certificate serial %s", req.Serial)
-	}
-
-	cert, err := SelectCertificate(ssa.dbMap.WithContext(ctx), req.Serial)
-	if db.IsNoRows(err) {
-		return nil, berrors.NotFoundError("certificate with serial %q not found", req.Serial)
-	}
-	if err != nil {
-		return nil, err
-	}
-	return bgrpc.CertToPB(cert), nil
-}
-
-// GetCertificateStatus takes a hexadecimal string representing the full 128-bit serial
-// number of a certificate and returns data about that certificate's current
-// validity.
-func (ssa *SQLStorageAuthority) GetCertificateStatus(ctx context.Context, req *sapb.Serial) (*corepb.CertificateStatus, error) {
-	if req.Serial == "" {
-		return nil, errIncompleteRequest
-	}
-	if !core.ValidSerial(req.Serial) {
-		err := fmt.Errorf("Invalid certificate serial %s", req.Serial)
-		return nil, err
-	}
-
-	certStatus, err := SelectCertificateStatus(ssa.dbMap.WithContext(ctx), req.Serial)
-	if db.IsNoRows(err) {
-		return nil, berrors.NotFoundError("certificate status with serial %q not found", req.Serial)
-	}
-	if err != nil {
-		return nil, err
-	}
-
-	return bgrpc.CertStatusToPB(certStatus), nil
-}
-
-// GetRevocationStatus takes a hexadecimal string representing the full serial
-// number of a certificate and returns a minimal set of data about that cert's
-// current validity.
-func (ssa *SQLStorageAuthority) GetRevocationStatus(ctx context.Context, req *sapb.Serial) (*sapb.RevocationStatus, error) {
-	if req.Serial == "" {
-		return nil, errIncompleteRequest
-	}
-	if !core.ValidSerial(req.Serial) {
-		return nil, fmt.Errorf("invalid certificate serial %s", req.Serial)
-	}
-
-	status, err := SelectRevocationStatus(ssa.dbMap.WithContext(ctx), req.Serial)
-	if err != nil {
-		if db.IsNoRows(err) {
-			return nil, berrors.NotFoundError("certificate status with serial %q not found", req.Serial)
-		}
-		return nil, err
-	}
-
-	return status, nil
+	return NewSQLStorageAuthorityWrapping(ssaro, dbMap, stats)
 }
 
 // NewRegistration stores a new Registration
@@ -606,254 +260,6 @@ func (ssa *SQLStorageAuthority) AddCertificate(ctx context.Context, req *sapb.Ad
 	}
 
 	return &sapb.AddCertificateResponse{Digest: digest}, nil
-}
-
-func (ssa *SQLStorageAuthority) CountOrders(ctx context.Context, req *sapb.CountOrdersRequest) (*sapb.Count, error) {
-	if req.AccountID == 0 || req.Range.Earliest == 0 || req.Range.Latest == 0 {
-		return nil, errIncompleteRequest
-	}
-
-	if features.Enabled(features.FasterNewOrdersRateLimit) {
-		return countNewOrders(ssa.dbReadOnlyMap.WithContext(ctx), req)
-	}
-
-	var count int64
-	err := ssa.dbReadOnlyMap.WithContext(ctx).SelectOne(
-		&count,
-		`SELECT count(1) FROM orders
-		WHERE registrationID = :acctID AND
-		created >= :earliest AND
-		created < :latest`,
-		map[string]interface{}{
-			"acctID":   req.AccountID,
-			"earliest": time.Unix(0, req.Range.Earliest),
-			"latest":   time.Unix(0, req.Range.Latest),
-		},
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	return &sapb.Count{Count: count}, nil
-}
-
-// HashNames returns a hash of the names requested. This is intended for use
-// when interacting with the orderFqdnSets table.
-func HashNames(names []string) []byte {
-	names = core.UniqueLowerNames(names)
-	hash := sha256.Sum256([]byte(strings.Join(names, ",")))
-	return hash[:]
-}
-
-func addFQDNSet(db db.Inserter, names []string, serial string, issued time.Time, expires time.Time) error {
-	return db.Insert(&core.FQDNSet{
-		SetHash: HashNames(names),
-		Serial:  serial,
-		Issued:  issued,
-		Expires: expires,
-	})
-}
-
-// addOrderFQDNSet creates a new OrderFQDNSet row using the provided
-// information. This function accepts a transaction so that the orderFqdnSet
-// addition can take place within the order addition transaction. The caller is
-// required to rollback the transaction if an error is returned.
-func addOrderFQDNSet(
-	db db.Inserter,
-	names []string,
-	orderID int64,
-	regID int64,
-	expires time.Time) error {
-	return db.Insert(&orderFQDNSet{
-		SetHash:        HashNames(names),
-		OrderID:        orderID,
-		RegistrationID: regID,
-		Expires:        expires,
-	})
-}
-
-// deleteOrderFQDNSet deletes a OrderFQDNSet row that matches the provided
-// orderID. This function accepts a transaction so that the deletion can
-// take place within the finalization transaction. The caller is required to
-// rollback the transaction if an error is returned.
-func deleteOrderFQDNSet(
-	db db.Execer,
-	orderID int64) error {
-
-	result, err := db.Exec(`
-	  DELETE FROM orderFqdnSets
-		WHERE orderID = ?`,
-		orderID)
-	if err != nil {
-		return err
-	}
-	rowsDeleted, err := result.RowsAffected()
-	if err != nil {
-		return err
-	}
-	// We always expect there to be an order FQDN set row for each
-	// pending/processing order that is being finalized. If there isn't one then
-	// something is amiss and should be raised as an internal server error
-	if rowsDeleted == 0 {
-		return berrors.InternalServerError("No orderFQDNSet exists to delete")
-	}
-	return nil
-}
-
-func addIssuedNames(db db.Execer, cert *x509.Certificate, isRenewal bool) error {
-	if len(cert.DNSNames) == 0 {
-		return berrors.InternalServerError("certificate has no DNSNames")
-	}
-	var qmarks []string
-	var values []interface{}
-	for _, name := range cert.DNSNames {
-		values = append(values,
-			ReverseName(name),
-			core.SerialToString(cert.SerialNumber),
-			cert.NotBefore,
-			isRenewal)
-		qmarks = append(qmarks, "(?, ?, ?, ?)")
-	}
-	query := `INSERT INTO issuedNames (reversedName, serial, notBefore, renewal) VALUES ` + strings.Join(qmarks, ", ") + `;`
-	_, err := db.Exec(query, values...)
-	return err
-}
-
-// CountFQDNSets counts the total number of issuances, for a set of domains,
-// that occurred during a given window of time.
-func (ssa *SQLStorageAuthority) CountFQDNSets(ctx context.Context, req *sapb.CountFQDNSetsRequest) (*sapb.Count, error) {
-	if req.Window == 0 || len(req.Domains) == 0 {
-		return nil, errIncompleteRequest
-	}
-
-	var count int64
-	err := ssa.dbReadOnlyMap.WithContext(ctx).SelectOne(
-		&count,
-		`SELECT COUNT(1) FROM fqdnSets
-		WHERE setHash = ?
-		AND issued > ?`,
-		HashNames(req.Domains),
-		ssa.clk.Now().Add(-time.Duration(req.Window)),
-	)
-	return &sapb.Count{Count: count}, err
-}
-
-// FQDNSetTimestampsForWindow returns the issuance timestamps for each
-// certificate, issued for a set of domains, during a given window of time,
-// starting from the most recent issuance.
-func (ssa *SQLStorageAuthority) FQDNSetTimestampsForWindow(ctx context.Context, req *sapb.CountFQDNSetsRequest) (*sapb.Timestamps, error) {
-	if req.Window == 0 || len(req.Domains) == 0 {
-		return nil, errIncompleteRequest
-	}
-	type row struct {
-		Issued time.Time
-	}
-	var rows []row
-	_, err := ssa.dbReadOnlyMap.WithContext(ctx).Select(
-		&rows,
-		`SELECT issued FROM fqdnSets 
-		WHERE setHash = ?
-		AND issued > ?
-		ORDER BY issued DESC`,
-		HashNames(req.Domains),
-		ssa.clk.Now().Add(-time.Duration(req.Window)),
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	var results []int64
-	for _, i := range rows {
-		results = append(results, i.Issued.UnixNano())
-	}
-	return &sapb.Timestamps{Timestamps: results}, nil
-}
-
-// FQDNSetExists returns a bool indicating if one or more FQDN sets |names|
-// exists in the database
-func (ssa *SQLStorageAuthority) FQDNSetExists(ctx context.Context, req *sapb.FQDNSetExistsRequest) (*sapb.Exists, error) {
-	if len(req.Domains) == 0 {
-		return nil, errIncompleteRequest
-	}
-	exists, err := ssa.checkFQDNSetExists(ssa.dbMap.WithContext(ctx).SelectOne, req.Domains)
-	if err != nil {
-		return nil, err
-	}
-	return &sapb.Exists{Exists: exists}, nil
-}
-
-// oneSelectorFunc is a func type that matches both gorp.Transaction.SelectOne
-// and gorp.DbMap.SelectOne.
-type oneSelectorFunc func(holder interface{}, query string, args ...interface{}) error
-
-// checkFQDNSetExists uses the given oneSelectorFunc to check whether an fqdnSet
-// for the given names exists.
-func (ssa *SQLStorageAuthority) checkFQDNSetExists(selector oneSelectorFunc, names []string) (bool, error) {
-	namehash := HashNames(names)
-	var exists bool
-	err := selector(
-		&exists,
-		`SELECT EXISTS (SELECT id FROM fqdnSets WHERE setHash = ? LIMIT 1)`,
-		namehash,
-	)
-	return exists, err
-}
-
-// PreviousCertificateExists returns true iff there was at least one certificate
-// issued with the provided domain name, and the most recent such certificate
-// was issued by the provided registration ID. This method is currently only
-// used to determine if a certificate has previously been issued for a given
-// domain name in order to determine if validations should be allowed during
-// the v1 API shutoff.
-// TODO(#5816): Consider removing this method, as it has no callers.
-func (ssa *SQLStorageAuthority) PreviousCertificateExists(ctx context.Context, req *sapb.PreviousCertificateExistsRequest) (*sapb.Exists, error) {
-	if req.Domain == "" || req.RegID == 0 {
-		return nil, errIncompleteRequest
-	}
-
-	exists := &sapb.Exists{Exists: true}
-	notExists := &sapb.Exists{Exists: false}
-
-	// Find the most recently issued certificate containing this domain name.
-	var serial string
-	err := ssa.dbMap.WithContext(ctx).SelectOne(
-		&serial,
-		`SELECT serial FROM issuedNames
-		WHERE reversedName = ?
-		ORDER BY notBefore DESC
-		LIMIT 1`,
-		ReverseName(req.Domain),
-	)
-	if err != nil {
-		if db.IsNoRows(err) {
-			return notExists, nil
-		}
-		return nil, err
-	}
-
-	// Check whether that certificate was issued to the specified account.
-	var count int
-	err = ssa.dbMap.WithContext(ctx).SelectOne(
-		&count,
-		`SELECT COUNT(1) FROM certificates
-		WHERE serial = ?
-		AND registrationID = ?`,
-		serial,
-		req.RegID,
-	)
-	if err != nil {
-		// If no rows found, that means the certificate we found in issuedNames wasn't
-		// issued by the registration ID we are checking right now, but is not an
-		// error.
-		if db.IsNoRows(err) {
-			return notExists, nil
-		}
-		return nil, err
-	}
-	if count > 0 {
-		return exists, nil
-	}
-	return notExists, nil
 }
 
 // DeactivateRegistration deactivates a currently valid registration
@@ -1237,321 +643,6 @@ func (ssa *SQLStorageAuthority) FinalizeOrder(ctx context.Context, req *sapb.Fin
 	return &emptypb.Empty{}, nil
 }
 
-// authzForOrder retrieves the authorization IDs for an order.
-func (ssa *SQLStorageAuthority) authzForOrder(ctx context.Context, orderID int64) ([]int64, error) {
-	var v2IDs []int64
-	_, err := ssa.dbMap.WithContext(ctx).Select(
-		&v2IDs,
-		"SELECT authzID FROM orderToAuthz2 WHERE orderID = ?",
-		orderID,
-	)
-	return v2IDs, err
-}
-
-// namesForOrder finds all of the requested names associated with an order. The
-// names are returned in their reversed form (see `sa.ReverseName`).
-func (ssa *SQLStorageAuthority) namesForOrder(ctx context.Context, orderID int64) ([]string, error) {
-	var reversedNames []string
-	_, err := ssa.dbMap.WithContext(ctx).Select(
-		&reversedNames,
-		`SELECT reversedName
-	   FROM requestedNames
-	   WHERE orderID = ?`,
-		orderID)
-	if err != nil {
-		return nil, err
-	}
-	return reversedNames, nil
-}
-
-// GetOrder is used to retrieve an already existing order object
-func (ssa *SQLStorageAuthority) GetOrder(ctx context.Context, req *sapb.OrderRequest) (*corepb.Order, error) {
-	if req == nil || req.Id == 0 {
-		return nil, errIncompleteRequest
-	}
-
-	omObj, err := ssa.dbMap.WithContext(ctx).Get(orderModel{}, req.Id)
-	if err != nil {
-		if db.IsNoRows(err) {
-			return nil, berrors.NotFoundError("no order found for ID %d", req.Id)
-		}
-		return nil, err
-	}
-	if omObj == nil {
-		return nil, berrors.NotFoundError("no order found for ID %d", req.Id)
-	}
-	order, err := modelToOrder(omObj.(*orderModel))
-	if err != nil {
-		return nil, err
-	}
-	orderExp := time.Unix(0, order.Expires)
-	if orderExp.Before(ssa.clk.Now()) {
-		return nil, berrors.NotFoundError("no order found for ID %d", req.Id)
-	}
-
-	v2AuthzIDs, err := ssa.authzForOrder(ctx, order.Id)
-	if err != nil {
-		return nil, err
-	}
-	order.V2Authorizations = v2AuthzIDs
-
-	names, err := ssa.namesForOrder(ctx, order.Id)
-	if err != nil {
-		return nil, err
-	}
-	// The requested names are stored reversed to improve indexing performance. We
-	// need to reverse the reversed names here before giving them back to the
-	// caller.
-	reversedNames := make([]string, len(names))
-	for i, n := range names {
-		reversedNames[i] = ReverseName(n)
-	}
-	order.Names = reversedNames
-
-	// Calculate the status for the order
-	status, err := ssa.statusForOrder(ctx, order)
-	if err != nil {
-		return nil, err
-	}
-	order.Status = status
-
-	return order, nil
-}
-
-// statusForOrder examines the status of a provided order's authorizations to
-// determine what the overall status of the order should be. In summary:
-//   - If the order has an error, the order is invalid
-//   - If any of the order's authorizations are in any state other than
-//     valid or pending, the order is invalid.
-//   - If any of the order's authorizations are pending, the order is pending.
-//   - If all of the order's authorizations are valid, and there is
-//     a certificate serial, the order is valid.
-//   - If all of the order's authorizations are valid, and we have began
-//     processing, but there is no certificate serial, the order is processing.
-//   - If all of the order's authorizations are valid, and we haven't begun
-//     processing, then the order is status ready.
-//
-// An error is returned for any other case.
-func (ssa *SQLStorageAuthority) statusForOrder(ctx context.Context, order *corepb.Order) (string, error) {
-	// Without any further work we know an order with an error is invalid
-	if order.Error != nil {
-		return string(core.StatusInvalid), nil
-	}
-
-	// If the order is expired the status is invalid and we don't need to get
-	// order authorizations. Its important to exit early in this case because an
-	// order that references an expired authorization will be itself have been
-	// expired (because we match the order expiry to the associated authz expiries
-	// in ra.NewOrder), and expired authorizations may be purged from the DB.
-	// Because of this purging fetching the authz's for an expired order may
-	// return fewer authz objects than expected, triggering a 500 error response.
-	orderExpiry := time.Unix(0, order.Expires)
-	if orderExpiry.Before(ssa.clk.Now()) {
-		return string(core.StatusInvalid), nil
-	}
-
-	// Get the full Authorization objects for the order
-	authzValidityInfo, err := ssa.getAuthorizationStatuses(ctx, order.V2Authorizations)
-	// If there was an error getting the authorizations, return it immediately
-	if err != nil {
-		return "", err
-	}
-
-	// If getAuthorizationStatuses returned a different number of authorization
-	// objects than the order's slice of authorization IDs something has gone
-	// wrong worth raising an internal error about.
-	if len(authzValidityInfo) != len(order.V2Authorizations) {
-		return "", berrors.InternalServerError(
-			"getAuthorizationStatuses returned the wrong number of authorization statuses "+
-				"(%d vs expected %d) for order %d",
-			len(authzValidityInfo), len(order.V2Authorizations), order.Id)
-	}
-
-	// Keep a count of the authorizations seen
-	pendingAuthzs := 0
-	validAuthzs := 0
-	otherAuthzs := 0
-	expiredAuthzs := 0
-
-	// Loop over each of the order's authorization objects to examine the authz status
-	for _, info := range authzValidityInfo {
-		switch core.AcmeStatus(info.Status) {
-		case core.StatusPending:
-			pendingAuthzs++
-		case core.StatusValid:
-			validAuthzs++
-		case core.StatusInvalid:
-			otherAuthzs++
-		case core.StatusDeactivated:
-			otherAuthzs++
-		case core.StatusRevoked:
-			otherAuthzs++
-		default:
-			return "", berrors.InternalServerError(
-				"Order is in an invalid state. Authz has invalid status %s",
-				info.Status)
-		}
-		if info.Expires.Before(ssa.clk.Now()) {
-			expiredAuthzs++
-		}
-	}
-
-	// An order is invalid if **any** of its authzs are invalid, deactivated,
-	// revoked, or expired, see https://tools.ietf.org/html/rfc8555#section-7.1.6
-	if otherAuthzs > 0 || expiredAuthzs > 0 {
-		return string(core.StatusInvalid), nil
-	}
-	// An order is pending if **any** of its authzs are pending
-	if pendingAuthzs > 0 {
-		return string(core.StatusPending), nil
-	}
-
-	// An order is fully authorized if it has valid authzs for each of the order
-	// names
-	fullyAuthorized := len(order.Names) == validAuthzs
-
-	// If the order isn't fully authorized we've encountered an internal error:
-	// Above we checked for any invalid or pending authzs and should have returned
-	// early. Somehow we made it this far but also don't have the correct number
-	// of valid authzs.
-	if !fullyAuthorized {
-		return "", berrors.InternalServerError(
-			"Order has the incorrect number of valid authorizations & no pending, " +
-				"deactivated or invalid authorizations")
-	}
-
-	// If the order is fully authorized and the certificate serial is set then the
-	// order is valid
-	if fullyAuthorized && order.CertificateSerial != "" {
-		return string(core.StatusValid), nil
-	}
-
-	// If the order is fully authorized, and we have began processing it, then the
-	// order is processing.
-	if fullyAuthorized && order.BeganProcessing {
-		return string(core.StatusProcessing), nil
-	}
-
-	if fullyAuthorized && !order.BeganProcessing {
-		return string(core.StatusReady), nil
-	}
-
-	return "", berrors.InternalServerError(
-		"Order %d is in an invalid state. No state known for this order's "+
-			"authorizations", order.Id)
-}
-
-type authzValidity struct {
-	Status  string
-	Expires time.Time
-}
-
-func (ssa *SQLStorageAuthority) getAuthorizationStatuses(ctx context.Context, ids []int64) ([]authzValidity, error) {
-	var qmarks []string
-	var params []interface{}
-	for _, id := range ids {
-		qmarks = append(qmarks, "?")
-		params = append(params, id)
-	}
-	var validityInfo []struct {
-		Status  uint8
-		Expires time.Time
-	}
-	_, err := ssa.dbMap.WithContext(ctx).Select(
-		&validityInfo,
-		fmt.Sprintf("SELECT status, expires FROM authz2 WHERE id IN (%s)", strings.Join(qmarks, ",")),
-		params...,
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	allAuthzValidity := make([]authzValidity, len(validityInfo))
-	for i, info := range validityInfo {
-		allAuthzValidity[i] = authzValidity{
-			Status:  string(uintToStatus[info.Status]),
-			Expires: info.Expires,
-		}
-	}
-	return allAuthzValidity, nil
-}
-
-// GetOrderForNames tries to find a **pending** or **ready** order with the
-// exact set of names requested, associated with the given accountID. Only
-// unexpired orders are considered. If no order meeting these requirements is
-// found a nil corepb.Order pointer is returned.
-func (ssa *SQLStorageAuthority) GetOrderForNames(
-	ctx context.Context,
-	req *sapb.GetOrderForNamesRequest) (*corepb.Order, error) {
-
-	if req.AcctID == 0 || len(req.Names) == 0 {
-		return nil, errIncompleteRequest
-	}
-
-	// Hash the names requested for lookup in the orderFqdnSets table
-	fqdnHash := HashNames(req.Names)
-
-	// Find a possibly-suitable order. We don't include the account ID or order
-	// status in this query because there's no index that includes those, so
-	// including them could require the DB to scan extra rows.
-	// Instead, we select one unexpired order that matches the fqdnSet. If
-	// that order doesn't match the account ID or status we need, just return
-	// nothing. We use `ORDER BY expires ASC` because the index on
-	// (setHash, expires) is in ASC order. DESC would be slightly nicer from a
-	// user experience perspective but would be slow when there are many entries
-	// to sort.
-	// This approach works fine because in most cases there's only one account
-	// issuing for a given name. If there are other accounts issuing for the same
-	// name, it just means order reuse happens less often.
-	var result struct {
-		OrderID        int64
-		RegistrationID int64
-	}
-	var err error
-	err = ssa.dbMap.WithContext(ctx).SelectOne(&result, `
-					SELECT orderID, registrationID
-					FROM orderFqdnSets
-					WHERE setHash = ?
-					AND expires > ?
-					ORDER BY expires ASC
-					LIMIT 1`,
-		fqdnHash, ssa.clk.Now())
-
-	if db.IsNoRows(err) {
-		return nil, berrors.NotFoundError("no order matching request found")
-	} else if err != nil {
-		return nil, err
-	}
-
-	if result.RegistrationID != req.AcctID {
-		return nil, berrors.NotFoundError("no order matching request found")
-	}
-
-	// Get the order
-	order, err := ssa.GetOrder(ctx, &sapb.OrderRequest{Id: result.OrderID})
-	if err != nil {
-		return nil, err
-	}
-	// Only return a pending or ready order
-	if order.Status != string(core.StatusPending) &&
-		order.Status != string(core.StatusReady) {
-		return nil, berrors.NotFoundError("no order matching request found")
-	}
-	return order, nil
-}
-
-func AuthzMapToPB(m map[string]*core.Authorization) (*sapb.Authorizations, error) {
-	resp := &sapb.Authorizations{}
-	for k, v := range m {
-		authzPB, err := bgrpc.AuthzToPB(*v)
-		if err != nil {
-			return nil, err
-		}
-		resp.Authz = append(resp.Authz, &sapb.Authorizations_MapElement{Domain: k, Authz: authzPB})
-	}
-	return resp, nil
-}
-
 // NewAuthorizations2 adds a set of new style authorizations to the database and
 // returns either the IDs of the authorizations or an error.
 // TODO(#5816): Consider removing this method, as it has no callers.
@@ -1619,103 +710,6 @@ func (ssa *SQLStorageAuthority) NewAuthorizations2(ctx context.Context, req *sap
 		return nil, err
 	}
 	return ids, nil
-}
-
-// GetAuthorization2 returns the authz2 style authorization identified by the provided ID or an error.
-// If no authorization is found matching the ID a berrors.NotFound type error is returned.
-func (ssa *SQLStorageAuthority) GetAuthorization2(ctx context.Context, req *sapb.AuthorizationID2) (*corepb.Authorization, error) {
-	if req.Id == 0 {
-		return nil, errIncompleteRequest
-	}
-	obj, err := ssa.dbMap.Get(authzModel{}, req.Id)
-	if err != nil {
-		return nil, err
-	}
-	if obj == nil {
-		return nil, berrors.NotFoundError("authorization %d not found", req.Id)
-	}
-	return modelToAuthzPB(*(obj.(*authzModel)))
-}
-
-// authzModelMapToPB converts a mapping of domain name to authzModels into a
-// protobuf authorizations map
-func authzModelMapToPB(m map[string]authzModel) (*sapb.Authorizations, error) {
-	resp := &sapb.Authorizations{}
-	for k, v := range m {
-		authzPB, err := modelToAuthzPB(v)
-		if err != nil {
-			return nil, err
-		}
-		resp.Authz = append(resp.Authz, &sapb.Authorizations_MapElement{Domain: k, Authz: authzPB})
-	}
-	return resp, nil
-}
-
-// GetAuthorizations2 returns any valid or pending authorizations that exist for the list of domains
-// provided. If both a valid and pending authorization exist only the valid one will be returned.
-func (ssa *SQLStorageAuthority) GetAuthorizations2(ctx context.Context, req *sapb.GetAuthorizationsRequest) (*sapb.Authorizations, error) {
-	if len(req.Domains) == 0 || req.RegistrationID == 0 || req.Now == 0 {
-		return nil, errIncompleteRequest
-	}
-	var authzModels []authzModel
-	params := []interface{}{
-		req.RegistrationID,
-		statusUint(core.StatusValid),
-		statusUint(core.StatusPending),
-		time.Unix(0, req.Now),
-		identifierTypeToUint[string(identifier.DNS)],
-	}
-
-	useIndex := ""
-	if features.Enabled(features.GetAuthzUseIndex) {
-		useIndex = "USE INDEX (regID_identifier_status_expires_idx)"
-	}
-
-	qmarks := make([]string, len(req.Domains))
-	for i, n := range req.Domains {
-		qmarks[i] = "?"
-		params = append(params, n)
-	}
-
-	query := fmt.Sprintf(
-		`SELECT %s FROM authz2
-			%s
-			WHERE registrationID = ? AND
-			status IN (?,?) AND
-			expires > ? AND
-			identifierType = ? AND
-			identifierValue IN (%s)`,
-		authzFields,
-		useIndex,
-		strings.Join(qmarks, ","),
-	)
-
-	dbMap := ssa.dbMap
-	if features.Enabled(features.GetAuthzReadOnly) {
-		dbMap = ssa.dbReadOnlyMap
-	}
-	_, err := dbMap.Select(
-		&authzModels,
-		query,
-		params...,
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(authzModels) == 0 {
-		return &sapb.Authorizations{}, nil
-	}
-
-	authzModelMap := make(map[string]authzModel)
-	for _, am := range authzModels {
-		existing, present := authzModelMap[am.IdentifierValue]
-		if !present || uintToStatus[existing.Status] == core.StatusPending && uintToStatus[am.Status] == core.StatusValid {
-			authzModelMap[am.IdentifierValue] = am
-		}
-	}
-
-	return authzModelMapToPB(authzModelMap)
 }
 
 // FinalizeAuthorization2 moves a pending authorization to either the valid or invalid status. If
@@ -1897,211 +891,6 @@ func (ssa *SQLStorageAuthority) UpdateRevokedCertificate(ctx context.Context, re
 	return &emptypb.Empty{}, nil
 }
 
-// GetPendingAuthorization2 returns the most recent Pending authorization with
-// the given identifier, if available. This method only supports DNS identifier types.
-// TODO(#5816): Consider removing this method, as it has no callers.
-func (ssa *SQLStorageAuthority) GetPendingAuthorization2(ctx context.Context, req *sapb.GetPendingAuthorizationRequest) (*corepb.Authorization, error) {
-	if req.RegistrationID == 0 || req.IdentifierValue == "" || req.ValidUntil == 0 {
-		return nil, errIncompleteRequest
-	}
-	var am authzModel
-	err := ssa.dbMap.WithContext(ctx).SelectOne(
-		&am,
-		fmt.Sprintf(`SELECT %s FROM authz2 WHERE
-			registrationID = :regID AND
-			status = :status AND
-			expires > :validUntil AND
-			identifierType = :dnsType AND
-			identifierValue = :ident
-			ORDER BY expires ASC
-			LIMIT 1 `, authzFields),
-		map[string]interface{}{
-			"regID":      req.RegistrationID,
-			"status":     statusUint(core.StatusPending),
-			"validUntil": time.Unix(0, req.ValidUntil),
-			"dnsType":    identifierTypeToUint[string(identifier.DNS)],
-			"ident":      req.IdentifierValue,
-		},
-	)
-	if err != nil {
-		if db.IsNoRows(err) {
-			return nil, berrors.NotFoundError("pending authz not found")
-		}
-		return nil, err
-	}
-	return modelToAuthzPB(am)
-}
-
-// CountPendingAuthorizations2 returns the number of pending, unexpired authorizations
-// for the given registration.
-func (ssa *SQLStorageAuthority) CountPendingAuthorizations2(ctx context.Context, req *sapb.RegistrationID) (*sapb.Count, error) {
-	if req.Id == 0 {
-		return nil, errIncompleteRequest
-	}
-
-	var count int64
-	err := ssa.dbReadOnlyMap.WithContext(ctx).SelectOne(&count,
-		`SELECT COUNT(1) FROM authz2 WHERE
-		registrationID = :regID AND
-		expires > :expires AND
-		status = :status`,
-		map[string]interface{}{
-			"regID":   req.Id,
-			"expires": ssa.clk.Now(),
-			"status":  statusUint(core.StatusPending),
-		},
-	)
-	if err != nil {
-		return nil, err
-	}
-	return &sapb.Count{Count: count}, nil
-}
-
-// GetValidOrderAuthorizations2 is used to find the valid, unexpired authorizations
-// associated with a specific order and account ID.
-func (ssa *SQLStorageAuthority) GetValidOrderAuthorizations2(ctx context.Context, req *sapb.GetValidOrderAuthorizationsRequest) (*sapb.Authorizations, error) {
-	if req.AcctID == 0 || req.Id == 0 {
-		return nil, errIncompleteRequest
-	}
-
-	var ams []authzModel
-	_, err := ssa.dbMap.WithContext(ctx).Select(
-		&ams,
-		fmt.Sprintf(`SELECT %s FROM authz2
-			LEFT JOIN orderToAuthz2 ON authz2.ID = orderToAuthz2.authzID
-			WHERE authz2.registrationID = :regID AND
-			authz2.expires > :expires AND
-			authz2.status = :status AND
-			orderToAuthz2.orderID = :orderID`,
-			authzFields,
-		),
-		map[string]interface{}{
-			"regID":   req.AcctID,
-			"expires": ssa.clk.Now(),
-			"status":  statusUint(core.StatusValid),
-			"orderID": req.Id,
-		},
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	byName := make(map[string]authzModel)
-	for _, am := range ams {
-		if uintToIdentifierType[am.IdentifierType] != string(identifier.DNS) {
-			return nil, fmt.Errorf("unknown identifier type: %q on authz id %d", am.IdentifierType, am.ID)
-		}
-		existing, present := byName[am.IdentifierValue]
-		if !present || am.Expires.After(existing.Expires) {
-			byName[am.IdentifierValue] = am
-		}
-	}
-
-	return authzModelMapToPB(byName)
-}
-
-// CountInvalidAuthorizations2 counts invalid authorizations for a user expiring
-// in a given time range. This method only supports DNS identifier types.
-func (ssa *SQLStorageAuthority) CountInvalidAuthorizations2(ctx context.Context, req *sapb.CountInvalidAuthorizationsRequest) (*sapb.Count, error) {
-	if req.RegistrationID == 0 || req.Hostname == "" || req.Range.Earliest == 0 || req.Range.Latest == 0 {
-		return nil, errIncompleteRequest
-	}
-
-	var count int64
-	err := ssa.dbReadOnlyMap.WithContext(ctx).SelectOne(
-		&count,
-		`SELECT COUNT(1) FROM authz2 WHERE
-		registrationID = :regID AND
-		status = :status AND
-		expires > :expiresEarliest AND
-		expires <= :expiresLatest AND
-		identifierType = :dnsType AND
-		identifierValue = :ident`,
-		map[string]interface{}{
-			"regID":           req.RegistrationID,
-			"dnsType":         identifierTypeToUint[string(identifier.DNS)],
-			"ident":           req.Hostname,
-			"expiresEarliest": time.Unix(0, req.Range.Earliest),
-			"expiresLatest":   time.Unix(0, req.Range.Latest),
-			"status":          statusUint(core.StatusInvalid),
-		},
-	)
-	if err != nil {
-		return nil, err
-	}
-	return &sapb.Count{Count: count}, nil
-}
-
-// GetValidAuthorizations2 returns the latest authorization for all
-// domain names that the account has authorizations for. This method is
-// intended to deprecate GetValidAuthorizations. This method only supports
-// DNS identifier types.
-func (ssa *SQLStorageAuthority) GetValidAuthorizations2(ctx context.Context, req *sapb.GetValidAuthorizationsRequest) (*sapb.Authorizations, error) {
-	if len(req.Domains) == 0 || req.RegistrationID == 0 || req.Now == 0 {
-		return nil, errIncompleteRequest
-	}
-
-	var authzModels []authzModel
-	params := []interface{}{
-		req.RegistrationID,
-		statusUint(core.StatusValid),
-		time.Unix(0, req.Now),
-		identifierTypeToUint[string(identifier.DNS)],
-	}
-	qmarks := make([]string, len(req.Domains))
-	for i, n := range req.Domains {
-		qmarks[i] = "?"
-		params = append(params, n)
-	}
-	_, err := ssa.dbMap.Select(
-		&authzModels,
-		fmt.Sprintf(
-			`SELECT %s FROM authz2 WHERE
-			registrationID = ? AND
-			status = ? AND
-			expires > ? AND
-			identifierType = ? AND
-			identifierValue IN (%s)`,
-			authzFields,
-			strings.Join(qmarks, ","),
-		),
-		params...,
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	authzMap := make(map[string]authzModel, len(authzModels))
-	for _, am := range authzModels {
-		// Only allow DNS identifiers
-		if uintToIdentifierType[am.IdentifierType] != string(identifier.DNS) {
-			continue
-		}
-		// If there is an existing authorization in the map only replace it with one
-		// which has a later expiry.
-		if existing, present := authzMap[am.IdentifierValue]; present && am.Expires.Before(existing.Expires) {
-			continue
-		}
-		authzMap[am.IdentifierValue] = am
-	}
-	return authzModelMapToPB(authzMap)
-}
-
-func addKeyHash(db db.Inserter, cert *x509.Certificate) error {
-	if cert.RawSubjectPublicKeyInfo == nil {
-		return errors.New("certificate has a nil RawSubjectPublicKeyInfo")
-	}
-	h := sha256.Sum256(cert.RawSubjectPublicKeyInfo)
-	khm := &keyHashModel{
-		KeyHash:      h[:],
-		CertNotAfter: cert.NotAfter,
-		CertSerial:   core.SerialToString(cert.SerialNumber),
-	}
-	return db.Insert(khm)
-}
-
-var blockedKeysColumns = "keyHash, added, source, comment"
-
 // AddBlockedKey adds a key hash to the blockedKeys table
 func (ssa *SQLStorageAuthority) AddBlockedKey(ctx context.Context, req *sapb.AddBlockedKeyRequest) (*emptypb.Empty, error) {
 	if core.IsAnyNilOrZero(req.KeyHash, req.Added, req.Source) {
@@ -2138,199 +927,6 @@ func (ssa *SQLStorageAuthority) AddBlockedKey(ctx context.Context, req *sapb.Add
 	return &emptypb.Empty{}, nil
 }
 
-// KeyBlocked checks if a key, indicated by a hash, is present in the blockedKeys table
-func (ssa *SQLStorageAuthority) KeyBlocked(ctx context.Context, req *sapb.KeyBlockedRequest) (*sapb.Exists, error) {
-	if req == nil || req.KeyHash == nil {
-		return nil, errIncompleteRequest
-	}
-
-	var id int64
-	err := ssa.dbMap.SelectOne(&id, `SELECT ID FROM blockedKeys WHERE keyHash = ?`, req.KeyHash)
-	if err != nil {
-		if db.IsNoRows(err) {
-			return &sapb.Exists{Exists: false}, nil
-		}
-		return nil, err
-	}
-
-	return &sapb.Exists{Exists: true}, nil
-}
-
-// IncidentsForSerial queries each active incident table and returns every
-// incident that currently impacts `req.Serial`.
-func (ssa *SQLStorageAuthority) IncidentsForSerial(ctx context.Context, req *sapb.Serial) (*sapb.Incidents, error) {
-	if req == nil {
-		return nil, errIncompleteRequest
-	}
-
-	var activeIncidents []incidentModel
-	_, err := ssa.dbMap.Select(&activeIncidents, `SELECT * FROM incidents WHERE enabled = 1`)
-	if err != nil {
-		if db.IsNoRows(err) {
-			return &sapb.Incidents{}, nil
-		}
-		return nil, err
-	}
-
-	var incidentsForSerial []*sapb.Incident
-	for _, i := range activeIncidents {
-		var count int
-		err := ssa.dbIncidentsMap.SelectOne(&count, fmt.Sprintf("SELECT COUNT(*) FROM %s WHERE serial = ?",
-			i.SerialTable), req.Serial)
-		if err != nil {
-			if db.IsNoRows(err) {
-				continue
-			}
-			return nil, err
-		}
-		if count > 0 {
-			incident := incidentModelToPB(i)
-			incidentsForSerial = append(incidentsForSerial, &incident)
-		}
-
-	}
-	if len(incidentsForSerial) == 0 {
-		return &sapb.Incidents{}, nil
-	}
-	return &sapb.Incidents{Incidents: incidentsForSerial}, nil
-}
-
-// SerialsForIncident queries the provided incident table and returns the
-// resulting rows as a stream of `*sapb.IncidentSerial`s. An `io.EOF` error
-// signals that there are no more serials to send. If the incident table in
-// question contains zero rows, only an `io.EOF` error is returned.
-func (ssa *SQLStorageAuthority) SerialsForIncident(req *sapb.SerialsForIncidentRequest, stream sapb.StorageAuthority_SerialsForIncidentServer) error {
-	if req.IncidentTable == "" {
-		return errIncompleteRequest
-	}
-
-	// Check that `req.IncidentTable` is a valid incident table name.
-	if !validIncidentTableRegexp.MatchString(req.IncidentTable) {
-		return fmt.Errorf("malformed table name %q", req.IncidentTable)
-	}
-
-	selector, err := db.NewMappedSelector[incidentSerialModel](ssa.dbIncidentsMap)
-	if err != nil {
-		return fmt.Errorf("initializing db map: %w", err)
-	}
-
-	rows, err := selector.QueryFrom(stream.Context(), req.IncidentTable, "")
-	if err != nil {
-		return fmt.Errorf("starting db query: %w", err)
-	}
-	defer rows.Close()
-
-	for rows.Next() {
-		// Scan the row into the model. Note: the fields must be passed in the
-		// same order as the columns returned by the query above.
-		ism, err := rows.Get()
-		if err != nil {
-			return err
-		}
-
-		err = stream.Send(
-			&sapb.IncidentSerial{
-				Serial:         ism.Serial,
-				RegistrationID: ism.RegistrationID,
-				OrderID:        ism.OrderID,
-				LastNoticeSent: ism.LastNoticeSent.UnixNano(),
-			})
-		if err != nil {
-			return err
-		}
-	}
-
-	err = rows.Err()
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-// GetRevokedCerts gets a request specifying an issuer and a period of time,
-// and writes to the output stream the set of all certificates issued by that
-// issuer which expire during that period of time and which have been revoked.
-// The starting timestamp is treated as inclusive (certs with exactly that
-// notAfter date are included), but the ending timestamp is exclusive (certs
-// with exactly that notAfter date are *not* included).
-func (ssa *SQLStorageAuthority) GetRevokedCerts(req *sapb.GetRevokedCertsRequest, stream sapb.StorageAuthority_GetRevokedCertsServer) error {
-	atTime := time.Unix(0, req.RevokedBefore)
-
-	clauses := `
-		WHERE notAfter >= ?
-		AND notAfter < ?
-		AND issuerID = ?
-		AND status = ?`
-	params := []interface{}{
-		time.Unix(0, req.ExpiresAfter),
-		time.Unix(0, req.ExpiresBefore),
-		req.IssuerNameID,
-		core.OCSPStatusRevoked,
-	}
-
-	selector, err := db.NewMappedSelector[crlEntryModel](ssa.dbReadOnlyMap)
-	if err != nil {
-		return fmt.Errorf("initializing db map: %w", err)
-	}
-
-	rows, err := selector.Query(stream.Context(), clauses, params...)
-	if err != nil {
-		return fmt.Errorf("reading db: %w", err)
-	}
-
-	defer func() {
-		err := rows.Close()
-		if err != nil {
-			ssa.log.AuditErrf("closing row reader: %s", err)
-		}
-	}()
-
-	for rows.Next() {
-		row, err := rows.Get()
-		if err != nil {
-			return fmt.Errorf("reading row: %w", err)
-		}
-
-		// Double-check that the cert wasn't revoked between the time at which we're
-		// constructing this snapshot CRL and right now. If the cert was revoked
-		// at-or-after the "atTime", we'll just include it in the next generation
-		// of CRLs.
-		if row.RevokedDate.After(atTime) || row.RevokedDate.Equal(atTime) {
-			continue
-		}
-
-		err = stream.Send(&corepb.CRLEntry{
-			Serial:    row.Serial,
-			Reason:    int32(row.RevokedReason),
-			RevokedAt: row.RevokedDate.UnixNano(),
-		})
-		if err != nil {
-			return fmt.Errorf("sending crl entry: %w", err)
-		}
-	}
-
-	err = rows.Err()
-	if err != nil {
-		return fmt.Errorf("iterating over row reader: %w", err)
-	}
-
-	return nil
-}
-
-// GetMaxExpiration returns the timestamp of the farthest-future notAfter date
-// found in the certificateStatus table. This provides an upper bound on how far
-// forward operations that need to cover all currently-unexpired certificates
-// have to look.
-func (ssa *SQLStorageAuthority) GetMaxExpiration(ctx context.Context, req *emptypb.Empty) (*timestamppb.Timestamp, error) {
-	var model struct {
-		MaxNotAfter time.Time `db:"maxNotAfter"`
-	}
-	err := ssa.dbReadOnlyMap.WithContext(ctx).SelectOne(
-		&model,
-		"SELECT MAX(notAfter) AS maxNotAfter FROM certificateStatus",
-	)
-	if err != nil {
-		return nil, fmt.Errorf("selecting max notAfter: %w", err)
-	}
-	return timestamppb.New(model.MaxNotAfter), err
-}
+// Note that the SQLStorageAuthority does have methods wrapping all of the
+// read-only methods provided by the SQLStorageAuthorityRO. These wrapper
+// implementations are in //sa/saro.go, next to the real implementations.
